backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
firebase-admin>=6.5.0
requests>=2.31.0
gunicorn>=22.0.0
whitenoise>=6.7.0
pytest-django>=4.8.0
